+++ backend/canisters/community/impl/src/lib.rs
@@ -10,7 +10,7 @@ use canister_timer_jobs::{Job, TimerJobs};
 use chat_events::{ChatEventInternal, ChatMetricsInternal};
 use community_canister::add_members_to_channel::UserFailedError;
 use community_canister::EventsResponse;
-use constants::{MINUTE_IN_MS, OPENCHAT_BOT_USER_ID, SNS_LEDGER_CANISTER_ID};
+use constants::{ICP_LEDGER_CANISTER_ID, MINUTE_IN_MS, OPENCHAT_BOT_USER_ID, SNS_LEDGER_CANISTER_ID};
 use event_store_producer::{EventStoreClient, EventStoreClientBuilder, EventStoreClientInfo};
 use event_store_producer_cdk_runtime::CdkRuntime;
 use fire_and_forget_handler::FireAndForgetHandler;
@@ -143,11 +143,12 @@ impl RuntimeState {
         let amount = treasury_share.saturating_sub(payment.fee);
         if amount > 0 {
             let is_chat = payment.ledger_canister_id == SNS_LEDGER_CANISTER_ID;
+            let is_icp = payment.ledger_canister_id == ICP_LEDGER_CANISTER_ID;
             self.data.pending_payments_queue.push(PendingPayment {
                 amount,
                 fee: if is_chat { 0 } else { payment.fee }, // No fee for BURNing
                 ledger_canister: payment.ledger_canister_id,
-                recipient: PaymentRecipient::Treasury,
+                recipient: if is_chat || is_icp { PaymentRecipient::SnsTreasury } else { PaymentRecipient::TreasuryCanister },
                 reason: PendingPaymentReason::AccessGate,
             });
         }