+++ backend/libraries/group_community_common/src/pending_payments_queue.rs
@@ -33,7 +33,8 @@ pub struct PendingPayment {
 
 #[derive(Serialize, Deserialize)]
 pub enum PaymentRecipient {
-    Treasury,
+    SnsTreasury,
+    TreasuryCanister,
     Member(UserId),
     Account(Account),
 }